moved dataswarm_task_table to ds_task_table
