+++ dataswarm/worker/Makefile
@@ -4,7 +4,7 @@ include ../../rules.mk
 TARGETS=$(PROGRAMS)
 PROGRAMS=dataswarm_worker
 
-SOURCES=dataswarm_worker_main.c dataswarm_worker.c dataswarm_task_table.c
+SOURCES=dataswarm_worker_main.c dataswarm_worker.c ds_task_table.c
 OBJECTS=$(SOURCES:%.c=%.o)
 DS_OBJECTS=../comm/ds_message.o ../task/ds_process.o ../task/ds_resources.o ../task/ds_task.o ../task/ds_mount.o ../blob/ds_blob.o ../blob/ds_blob_table.o
