+++ dataswarm/worker/ds_task_table.c
@@ -1,4 +1,4 @@
-#include "dataswarm_task_table.h"
+#include "ds_task_table.h"
 #include "dataswarm_worker.h"
 #include "task/ds_task.h"
 #include "comm/ds_message.h"
